% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{add_word_to_dictionary}
\alias{add_word_to_dictionary}
\title{add_word_to_dictionary}
\usage{
add_word_to_dictionary(new_word = NULL, word_class = NULL)
}
\arguments{
\item{new_word}{Det nye ord}

\item{word_class}{Det nye ords ordklasse}
}
\value{
Intet. Men gemmer ordbogen som datasæt i /data
}
\description{
Funktionen tilføjer et nyt ord til ordbogen
}
