% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{extract_comma}
\alias{extract_comma}
\title{extract comma}
\usage{
extract_comma(input_data, label)
}
\arguments{
\item{input_data}{data.frame}

\item{label}{Relevant feature (character-class)}
}
\value{
input_data plus 2 nye features: antal komma og komma_position. I
        \code{label}-kolonnen er kommaer fjernet.
}
\description{
Funktionen tæller antallet af kommaer i \code{label} i
             \code{input_data}, og danner en ny feature \code{antal_komma}.
             Derefter identificeres positionen af ordet før kommaet og
             returneres i featuren \code{komma_position}. Til sidst fjernes
             kommaer i \code{label}. \cr
             Hvis der i sætningen indgår semikolon, så opfattes det som
             et komma.
}
