% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{fill_3d_array}
\alias{fill_3d_array}
\title{fill_3d_array}
\usage{
fill_3d_array(sentence_list, pad_len, features)
}
\arguments{
\item{sentence_list}{Liste med character vectore.}

\item{pad_len}{Padding længde}

\item{features}{character vector med navne på de featuers der indgår}
}
\value{
Et 3d array med dimensionen {length(sentence_list), pad_len,
        length(features)}
}
\description{
Funktionen mapper inputlisten \code{sentence_list} til et 3d
             array hvor index [1,,] angiver
}
