% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{fit_model}
\usage{
fit_model(
  word_data,
  word_class_data,
  comma_list_index,
  comma_list_number,
  model_nyt = ""
)
}
\arguments{
\item{word_data}{...}

\item{word_class_data}{...}

\item{comma_list_index}{...}

\item{comma_list_number}{...}
}
\value{
...
}
\description{
fit_model
}
