% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_input.R
\name{format_input}
\alias{format_input}
\title{format_input}
\usage{
format_input(input_str, numCores = 4L)
}
\arguments{
\item{input_str}{...}

\item{numCores}{...}
}
\value{

}
\description{
format_input
}
