% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(path, file_pattern = NULL, gsub_pattern = "")
}
\arguments{
\item{path}{En character vector med stien til det data der skal hentes}

\item{file_pattern}{Et valgfrit regex udtryk til at hente specifikke datasæt}

\item{gsub_pattern}{Endelsen på de hentede filer. Eksemelvis ".rda"}
}
\value{
Data fra den specifirede sti i en liste.
}
\description{
Funktionen henter r-datasæt fra den specificerede \code{path}
             og laver en \code{\link[dplyr]{bind_rows}}.
}
