% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{gsub_custom}
\alias{gsub_custom}
\title{gsub custom}
\usage{
gsub_custom(string, replace)
}
\arguments{
\item{replace}{En named vektor med mønstre der skal erstattes}

\item{streng}{En tekststreng}
}
\value{
Input streng hvor gamle mønstre er udbyttet med de nye specificerede
        mønstre
}
\description{
Funktionen er en wrapper rundt om \code{base::gsub}, blot med
             en forlykke uden om. Input er en teststreng hvor en eller flere
             mønstre ønskes byttet ud. De mønstre der ønskes erstattet,
             skal speficeres i en vector som:
             vek <- c("nuværende mønster 1" = "nyt mønster 1",
             "nuværende mønster 2" = "nyt mønster 2").
}
