% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{handle_uppercase}
\alias{handle_uppercase}
\title{handle_uppercase}
\usage{
handle_uppercase(str_list, ord, navneordbog, impute_value)
}
\arguments{
\item{str_list}{Liste med sætninger hvor hver sætning er splitter op i ord}

\item{ord}{Ordbog som character vector}

\item{navneordbog}{Navne som character vektor}

\item{impute_value}{Værdier som skal imputes hvis uppercase ordet er ukendt}
}
\value{
\code{str_list} hvor uppercase er håndteret som ovenstående
        beskrivelse
}
\description{
Hvis et ord uppercase, med længde større en 1, ikke fremgår i ordbogen, men
fremgår i ordbogen som lowercase, så rettes ordet til lowercase. Hvis ordet
ikke fremgår i ordbogen, så tjekkes navneordbogen, hvis ordet heller ikke
findes her, så bliver ordet markeret som den indsatte værdi for
\code{impute_value}.
}
\note{
Ideen med at give ukendte uppercase ord sin egen klasse er, at disse
      ord har en anden karakteristik end almindelige ukendte ord. Tanken er,
      er disse ukendte uppercase ord typisk vil være forkortelser som
      eksempelvis GDPR eller FIFA.
      Længde større end 1 gør, at "I" ikke forvekles med "i".
}
