% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{identify_abbreviation}
\alias{identify_abbreviation}
\title{Identify abbrevation}
\usage{
identify_abbreviation(string, output_lenght = 10)
}
\arguments{
\item{string}{En tekststreng.}

\item{output_length}{Længden på strengen der printes til konsolen omkring
den fundne forkortelse.}
}
\value{
Tekst der opfylder beskrivelsen i description.
}
\description{
Funktionen undersøger om der i tekst-stregnen er forkortelser,
             som eksempelvis "bl.a." eller "fx.". Det gøres ved at
             identificere steder i strengen som har et stoptegn (punktum,
             spørgsmålstegn, udråbstegn, kolon) og det efterfølgende tegn
             er et space efterfuldt af et lowercase bogstav.
}
