% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{identify_number_before_punctuation}
\alias{identify_number_before_punctuation}
\title{Identificer tekst hvor et tal efterfølges af et sluttegn.}
\usage{
identify_number_before_punctuation(streng, output_lenght = 10)
}
\arguments{
\item{streng}{En tekststreng.}

\item{output_lenght}{Længden på strengen der printes til konsolen omkring
den fundne forkortelse.}
}
\value{
Tekst der opfylder beskrivelsen i description.
}
\description{
Identificer tekst hvor er tal efterfølges af et sluttegn.
}
