% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{kontruer_ordklasse}
\alias{kontruer_ordklasse}
\title{kontruer ordklasse}
\usage{
kontruer_ordklasse()
}
\description{
Her identificeres ordklassen til de ord der fremgår af
             ordbogen. Nogle ord har flere klasser, da det ikke er
             muligt fra gang til gang at tage stilling til den
             korrekte ordklasse, så vælges den ordklasse som fremgår flest
             gange først grupperet på ordet og derefter grupperet på
             ordklassen i hele datasættet. For dem hvor ingen ordklasse er
             tilknyttet gives værdien "ingen".
             Det konstruerede data gemmes i mappen /data under navnet
             "ord_og_klasse".
}
