% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{map_sentence_to_word_class}
\alias{map_sentence_to_word_class}
\title{map_sentence_to_dictionary_index}
\usage{
map_sentence_to_word_class(sentence, dictionary)
}
\arguments{
\item{sentence}{En character vector med sætninger}

\item{dictionary}{En dataframe med 2 kolonner; Ord og Klasse}

\item{split}{En character vector med regular expression}

\item{missing}{Inputationsværdi for manglende værdier}
}
\value{
En liste hvor hvert element er en numerisk vector

En liste hvor hvert element er en character vector
}
\description{
Funktionen mapper hvert ord i \code{sentence} til det korrekte
             index i \code{dictionary}. Det gøres ved at opslitte hver
             sætning efter regular expression udtrykket i \code{split}.
             Hvis et ord ikke findes i \code{dictionary}, så indsættes
             værdien defineret i \code{missing}.

... bruger \code{\link{split_sentence_to_word_regex}}
}
