% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{match_words_in_vocab}
\alias{match_words_in_vocab}
\title{match_words_in_vocab}
\usage{
match_words_in_vocab(str_list, ord, numCores)
}
\arguments{
\item{str_list}{Liste med sætninger hvor hver sætning er splitter op i ord}

\item{ord}{Ordbog som character vector}

\item{numCores}{Antal kerner}
}
\value{
En liste som \code{str_list}, men i stedet for ord, så indeholdende
        index ift. hvor ordet er matchet i \code{ord}.
}
\description{
Her matches hvert ord i hvert element i \code{str_list} i
\code{ord} og får det tilsvarende index nummer i \code{ord}.
}
\details{
I både \code{str_list} og \code{ord} findes ord hvor der kan indgå både store
og små bogstaver. For at matche så korrekt som muligt, bruges
hverken tolower() eller toupper() funktionalitet. Det har dog den
ulempe, at det første ord i hver sætning, som altid har stort
begyndelsesbogstav, sjældent matches i ordbogen. Derfor bruges
tolower() på første ord i hver sætning, og det første ord matches
både på det originale ord, med stort begyndelsesbogstav og på
tolower udgaven. Efterfølgende vælges det ord, som kunne matches i
\code{ord}. Hvis ingen kunne matches returneres *NA* for første ord, hvis
de begge kunne matches, så returneres tolower indekset. Det vurderes at det
er mest korrekt, selvom der kan være tilfælde, hvor det vil være mest korrekt
at returnere indekset for det oprindelige ord (dvs. uden tolower). Årsagen
til, at det er nødvendigt at matche det første ord 2 gange er, at et ord som
eksempelvis *Hvem* kun findes i tolower-udgaven i \code{ord}, mens et ord som
*USA* ikke findes i tolower-udgaven.
}
