% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{missing_data}
\alias{missing_data}
\title{missing_data}
\usage{
missing_data(data_splitted, indices_data)
}
\arguments{
\item{data_splitted}{En liste hvor hvert element er en character vektor
indeholdende en række ord}

\item{indices_data}{En liste hvor hvert element matcher det tilsvarende
element i \code{data_splitted} med samme længde.}
}
\value{
En data.frame med de ord, som ikke kunne matches og antallet af
        gange de optræder.
}
\description{
Funktionens formål er at identificere ord i
             \code{data_splitted} som ikke kunne matches i
             \code{\link{konstruer_ordbog}}. Ord der ikke kunne matches er
             angivet med \code{NA} i \code{indices_data}.
}
