% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{remove_additional_whitespaces}
\alias{remove_additional_whitespaces}
\title{Remove additional whitespaces}
\usage{
remove_additional_whitespaces(string)
}
\arguments{
\item{string}{En tekststreng}
}
\value{
Input strengen hvor overflødige mellemrum er fjernet.
}
\description{
Hvis der i en streng fremgår mere end 1 whitespace i træk, så
             fjernes de. Denne funktion fjerner også linjeskift. Derudover
             fjernes whitespaces før og efter strengen.
}
