% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{replace_na_with_regex}
\alias{replace_na_with_regex}
\title{replace_na_with_regex}
\usage{
replace_na_with_regex(match_data, splitted_data, regex_expr, replace_value)
}
\arguments{
\item{replace_value}{}
}
\value{
En liste med samme værdier som \code{splitted_data}, hvor NA er
        erstattet med værdien \code{replace_value} ... .
}
\description{
replace_na_with_regex
}
