% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{replace_na_with_vector}
\alias{replace_na_with_vector}
\title{replace na med vector}
\usage{
replace_na_with_vector(indices_data, data_splitted, word_vector, replace_value)
}
\arguments{
\item{replace_value}{}
}
\value{
En liste med samme værdier som \code{data_splitted}, hvor NA er
        erstattet med værdien \code{replace_value} ... .
}
\description{
...
}
