% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{sentence_in_parentheses}
\alias{sentence_in_parentheses}
\title{sentence_in_parentheses}
\usage{
sentence_in_parentheses(string)
}
\arguments{
\item{string}{En character vector}
}
\value{
En character-vector identisk med \code{string}, men med et indsat
        punktum før en parantes, hvis parantesen indeholder flere sætninger
        (defineret ved at punktum, udråbstegn eller spørgsmålstegn indgår).
}
\description{
Hvis der i \code{string} indgår en parentes, hvor der inde i
             parentesen indgår seperate sætninger (adskilt med ., ! eller ?)
             så vil der før parentesen starter, indsættes et punktum.
}
\details{
Hvis der inde i sætningen indgår en forkortelse, så vil den også
         blive splittet.
}
