% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{split_end_punctuation}
\alias{split_end_punctuation}
\title{split end punctuation}
\usage{
split_end_punctuation(char_vec)
}
\arguments{
\item{string}{En character vektor}
}
\value{
En matrix med kolonnerne sentence og end_punctuation
}
\description{
Funktionen tager input strengen og opslitter i sætning og
             sluttegn.
}
