% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{split_sentence_to_word_regex}
\alias{split_sentence_to_word_regex}
\title{split_sentence_to_word_regex}
\usage{
split_sentence_to_word_regex()
}
\value{
regex udtryk
}
\description{
Funktionen bruges blot til at returnere et regex udtryk som
             bruges til at splitte en sætning til enkeltstående ord.
             Det gøres flere steder i koden, så for ikke at skulle rette
             det flere steder, så laves det som en funktion.
}
\details{
Forklaring af de forskellige regex blokker:
    \itemize{
      \item blok 1: Splitter på space, komma, start- og slutparantes
      \item blok 2: Splitter på apostrof, betinget på at der enten er et
                    mellemrum før, eller et mellemrum efter (fordi nogle ord
                    eller navne naturligt slutter med apostrof og de skal
                    ikke splittes). Splitter også hvis apostrofen er i starten
                    eller i slutningen af sætningen, og i de tilfælde hvor der
                    er et komma eller en parentes efterfølgende.
      \item blok 3: splitter på komma, betinget på, at der efter et komma
                    skal være et space (så et decimaltal som 23,32 ikke bliver
                    splittet).
    }
}
