% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{split_str_to_dataframe}
\alias{split_str_to_dataframe}
\title{split_str_to_dataframe}
\usage{
split_str_to_dataframe(string, abbreviation = NULL)
}
\arguments{
\item{string}{Input streng}

\item{abbreviation}{Forkortelser som fx. "mr.", som efterfølges af et stort
start bogstav. Det parses ind i regex expression, så
sætningen ikke opdeles på disse forkortelser. For hver
input forkortelse indsættes et white space før
forkortelsen, således at et afsluttende ord i en sætning
som tilfældigvis ender på forkortelsen, ikke bliver
opdelt. Ligeledes indsættes et ^ (regex anchor), således
at sætningen ikke splittes, selvom forkortelsen er i
starten af sætningen.}
}
\value{
En data.frame med en varible, hvor hver række indeholder en sætning.
}
\description{
Funktionen opdeler i sætninger og tager delimiteren med. Her betinges på at
delimiteren skal være efterfulgt af mindst ét white space eller linjeskifte
\code{\n} samt at det er efterfulgt af et stort startbogstav eller tal.
Dette for at undgå, at en sætning ikke bliver splittet på en forkortelse.
}
