% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{swap_to_capital}
\alias{swap_to_capital}
\title{Swap to capital}
\usage{
swap_to_capital(string, match_symbol = NULL)
}
\arguments{
\item{match_symbol}{}

\item{streng}{En tekststreng.}
}
\value{
Inputstrengen med stort begyndelsesbogstav efter ? og !.
}
\description{
Hvis der efter et "?", "!", ":" (eller andet, kommer an på
             \code{streng}) er et lille begyndelsesbogstav, så rettes det til
             det tilsvarende store begyndelsesbogstav. I søge-strengen
             betinges på, at der skal være et ikke whitespace symbol før,
             og at der skal være et, eller flere, mellemrum eller
             linjeskifte \code{(\n)} efter.
}
