% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{tjek_data}
\alias{tjek_data}
\title{tjek_data}
\usage{
tjek_data(sentence_list, comma_index_list)
}
\arguments{
\item{sentence_list}{Liste med sætninger}

\item{comma_index_list}{Liste med komma index}
}
\value{
Funktionen returnerer intet
}
\description{
Funktionen tjekker om længden på hvert element i sentence_list
             er den samme som det tilsvarende element i comma_index_list.
}
