% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_funktioner.R
\name{vectorize_sequences}
\alias{vectorize_sequences}
\title{vectorize sequences}
\usage{
vectorize_sequences(sequence, dimension)
}
\arguments{
\item{sequence}{...}

\item{dimension}{...}
}
\value{
...
}
\description{
Mapper til binær matrice
}
\note{
Denne funktion er ikke testet!
}
